Introduce Characterization Test for Matlab results
To assure correct results, it would be nice to have an overarching
characterization test to guarantee obtaining the same results while
making changes to the codebase. It also helps a first implementation
of the Python code itself.

In this work, a general approach and structure is imposed to easily add
more input/output files to the Characterization test.

Temporarily, the test has been slimed by providing the expected output
directly, as the calculation logic is still to be implemented.