Remove lrmodules from root directory
At the moment, there are two `lrmodules` folders/packages:

* one in the project root
* one under `packages/`

As agreed previously, `lrmodule` should reside within the `packages`
directory, which was specifically created to manage mono-repos using UV
workspaces.

It’s unclear whether the duplicate was introduced due to lingering
commits missed during a Pull Request merge/rebase or if it was
intentional. In either case, the duplicate will be removed.

If there are features introduced in this duplicate package that need to
be preserved, please cherry-pick the relevant commit(s), move them into
the appropriate folder, perform a `module pull --rebase` against `main`
to ensure the latest changes, and open a new Pull Request.

Additionally, one of the CI quality tests recently started failing,
which prevented detection of certain code quality issues. Because this
project is private, it’s not possible to enforce automated checks. The
team is encouraged to follow best practices—avoid merging when tests are
failing, and communicate when CI tests are not functioning as expected.

Thank you for helping keep the codebase clean and consistent.

---

Please make commits make sense

Either rebase your commits before merging or squash reword your commit
and then merge. What is contributor suppose to understand in the future
about commits such as '.', 'rebase', or 'docstrings'

---

Add some code quality edits to allow tests to pass

- SHA1 is has security risks switch to SHA256
- Add some filler docstrings
- Overall ruff formatting