Rewrite tests with the railway way containers in mind
We no longer are working with the raw value, since it's unclear at
runtime if a failure occured or not. Thus the raw value will be
encapsulated in a container that can be passed around until the entended
reciever consumes it; think of a mail currier package.

State of the container can be Success or Failure, but that can be unpack
lazily